+++ module2/admin.py
@@ -9,6 +9,9 @@ class Module2Admin(admin.ModelAdmin):
     readonly_fields = []
 
     def get_readonly_fields(self, request, obj=None):
-        return list(self.readonly_fields) + \
-               [field.name for field in obj._meta.fields] + \
-               [field.name for field in obj._meta.many_to_many]
+        if len(self.readonly_fields) > 0:
+            return list(self.readonly_fields) + \
+                   [field.name for field in obj._meta.fields] + \
+                   [field.name for field in obj._meta.many_to_many]
+        else:
+            return list()