+++ docs/CHANGELOG.md
@@ -1,3 +1,18 @@
+# 3.0.0
+* Add access limits
+  * Switch from `login_required` to custom decorator `active_user_required` which checks whether the user has exceeded the limits
+  * Schema updates to add the override, counter and max limit
+  * Add util method to update the view counter
+  * Add the `limit_reached` page
+  * Redirect to `limit_reached` on home page if limits have been reached
+* Fix issue with module 2 and 3 admins not rendering the add/edit forms
+* Per Cheryl, display the restart link regardless of status
+
+# 2.1.0
+* Disable module 4 and STRIPE (not used)
+* Display TOU before showing the list of modules
+* Add `has_tou` flag in `area_app_user`
+
 # 2.0.0
 * Release module 3